% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_measures.R
\name{calc_chi_u_by_dist}
\alias{calc_chi_u_by_dist}
\title{Calculate chi_u for various distances (binned)}
\usage{
calc_chi_u_by_dist(X, coords, nbin, ps, bins = NULL)
}
\arguments{
\item{X}{matrix of data (columns: locations, rows: years)}

\item{coords}{n-location x 2 dimensional matrix of observation coordinates,
corresponding to the columns of X.}

\item{nbin}{number of equally spaced bins to use}

\item{ps}{vector of quantiles over which to estimate chi_u}

\item{bins}{(optional) alternatively, can explicitly pass a vector of bin
boundaries}
}
\value{
empirical estimate of chi_u for various binned distances
}
\description{
Calculate chi_u for various distances (binned)
}
