% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_mcmc.R
\name{calc_chi_u_by_dist_mcmc}
\alias{calc_chi_u_by_dist_mcmc}
\title{Posterior predictive estimat of chi_u as a function of distance}
\usage{
calc_chi_u_by_dist_mcmc(mh, coords, nbin, ps, type = c("smith", "br"),
  knot_coord = NULL, bins = NULL, thin_int = 1, n.core = NULL)
}
\arguments{
\item{mh}{MCMC samples from \code{mcmc_lnorm_basis} or
\code{mcmc_fixed_basis} output.}

\item{coords}{n-location x 2 dimensional matrix of observation coordinates}

\item{nbin}{number of bins to use (unique distances) when estimating chi_u}

\item{ps}{vector of probabilities (u terms in chi_u)}

\item{type}{one of "smith" or "br" for Gaussian density or log-Gaussian
process basis respectively.}

\item{knot_coord}{n-knot x 2 matrix of knot coordinates (needed only if
\code{type} = "smith")}

\item{bins}{(optional) vector of distances defining bin boundaries. If this
is not supplied, bins are defined between shortest distances and greatest
distance between observation locations.}

\item{thin_int}{(optional) thinning interval to use.}

\item{n.core}{number of cores to parallelize calculation over.}
}
\value{
Posterior predictive estimate of chi_u for different bin distances,
and quantiles defined by \code{ps}.
}
\description{
Posterior predictive estimat of chi_u as a function of distance
}
