% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{calc_dist}
\alias{calc_dist}
\title{Calcuate pairwise Euclidean distances between two matrices with d-dim columns}
\usage{
calc_dist(A, B)
}
\arguments{
\item{A}{Matrix 1 (nrow = m)}

\item{B}{Matrix 2 (nrow = n)}
}
\value{
An m by n matrix of distances between the rows of A and rows of B,
 where m = \code{nrow(x)} and n = \code{nrow(y)}
}
\description{
Calculate pairwise Euclidean distances between all rows of
a matrix A and all rows of another matrix B. The output matrix will be of dimension
}
\examples{
A1 <- matrix(1:6, ncol = 2)
B1 <- matrix(1:8, ncol = 2)
calc_dist(A1, B1)

}
