% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_measures.R
\name{calc_jpexcd}
\alias{calc_jpexcd}
\title{Calculate joint exceedance probability for all pairs listed in pair_ids}
\usage{
calc_jpexcd(X, pair_ids, p)
}
\arguments{
\item{X}{Matrix where data replicates are stored in separate columns}

\item{pair_ids}{2-column matrix containing column indicies of X matrix. The
function \code{calc_jpexcd_one_pair} will be applied to each pair of columns.}

\item{p}{upper quantile for which to estimate joint exceedance probability}
}
\value{
Mean joint exceedance probability estimate of exceeding pth quantile
as estimated by comparing all pairs of observation vectors listed in
\code{pair_ids} matrix
}
\description{
Calculate joint exceedance probability for all pairs listed in pair_ids
}
