% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_measures.R
\name{calc_jpexcd_one_pair}
\alias{calc_jpexcd_one_pair}
\title{Empirical estimate of joint exceedance probabilities}
\usage{
calc_jpexcd_one_pair(x, y, p)
}
\arguments{
\item{x}{first vector of observations}

\item{y}{second vector of observations}

\item{p}{upper quantile for which to estimate joint exceedance probability}
}
\value{
empirical estimate of joint exceedance probability.
}
\description{
Calculate joint exceedance probability for one pair of vectors
}
