% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{calc_log_scores}
\alias{calc_log_scores}
\title{Calculate log-scores}
\usage{
calc_log_scores(yhold, alpha_smps, theta_smps, lB_smps, lK_smps, mu_smps,
  sigma_smps, xi_smps)
}
\arguments{
\item{yhold}{GEV scale data on holdout set}

\item{alpha_smps}{posterior samples of alpha (positive stable index)}

\item{theta_smps}{posterior samples of theta (Exponentail tilting par.)}

\item{lB_smps}{posterior samples of log-transformed basis scaling factors
(each row is a different MCMC sample, and the columns contain the C style
column-major ordering when transforming a matrix to a vector of lB as defined
in for example rstabmix.)}

\item{lK_smps}{posterior samples of log-transformed basis functions at
holdout locations. See \code{mcmc_lnorm_basis} for dimensions of array.}

\item{mu_smps}{posterior samples of GEV location parameters
holdout locations. Rows: mcmc samples. Columns: holdout locations}

\item{sigma_smps}{posterior samples of GEV scale parameters
holdout locations. Rows: mcmc samples. Columns: holdout locations}

\item{xi_smps}{posterior samples of GEV shape parameters
holdout locations. Rows: mcmc samples. Columns: holdout locations}
}
\value{
log-scores for holdout set.
}
\description{
Calculate log-scores
}
