% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{calc_mpar}
\alias{calc_mpar}
\title{Add mean function to mean-zero GP}
\usage{
calc_mpar(X, beta, gp, nrep, nloc, log.link = FALSE)
}
\arguments{
\item{X}{matrix of known covariates used in mean function}

\item{beta}{linear coefficients for X matrix}

\item{gp}{mean zero gaussian process}

\item{nrep}{number of data replicates of stable mixture process that this GP
will later interface with}

\item{nloc}{number of spatial locations}

\item{log.link}{(logical) should a log link be used?}
}
\description{
Combine mean function with a mean-zero Gaussian process
}
