% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{chol_prop_lK}
\alias{chol_prop_lK}
\title{Block Cholesky de-noising proposals for lK (see paper appendix for details)}
\usage{
chol_prop_lK(lK_sub_cur, Csub, lK_prop_var_sub, nbasis_per_rep)
}
\arguments{
\item{lK_sub_cur}{Current values of log-Gaussian process basis functions at
a subset of the observation locations}

\item{Csub}{Covariance matrix of the Gaussian process at subset of obs
locations in \code{lK_sub_cur}}

\item{lK_prop_var_sub}{proposal variance for random walk}

\item{nbasis_per_rep}{(not currently used)}
}
\value{
lK proposal on region subset
}
\description{
Block Cholesky de-noising proposals for lK (see paper appendix for details)
}
