% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_simulation.R
\name{cond_sim_lK}
\alias{cond_sim_lK}
\title{Conditionally simulate Gaussian process basis functions}
\usage{
cond_sim_lK(lK, snew, sold, gvar, gscl)
}
\arguments{
\item{lK}{normalized log-Gaussian process basis function matrix on which to
condition observed at \code{sold} and transformed to log-scale}

\item{snew}{n-new-location x 2 dimensional coordinate matrix of locations at
which to conditionally simulate lK.}

\item{sold}{n-obs-location x 2 dimensional coordinate matrix of locations at
which samples of lK are to be conditioned on}

\item{gvar}{Gaussian process exponential covariance function variance}

\item{gscl}{Gassian process exponential covariance function scale/range}
}
\description{
Conditionally simulate Gaussian process basis functions
}
