% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{create_qlstabmix}
\alias{create_qlstabmix}
\title{Create list of quantile functions}
\usage{
create_qlstabmix(alpha, theta, lK)
}
\arguments{
\item{alpha}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{theta}{stable theta parameter (exponential tilting parameter). This
parameter controls the tail dependence class. \code{theta} = 0 gives
asymptotic dependence, and \code{theta} > 0 gives asymptotic independence.}

\item{lK}{log-kernel basis matrix (rows: locations, columns: different
kernel functions)}
}
\value{
List of log-stable mixture observation model quantile functions
}
\description{
Create list of quantile functions for log-stablemix data model
elements of the list correspond to a single observation location. List is in
same order as rows of log-basis matrix \code{lK}
}
