% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{dhexpstabR}
\alias{dhexpstabR}
\title{R implementation of density function for exponentially tilted PS(alpha)}
\usage{
dhexpstabR(x, alpha, delta, theta, return_log = TRUE, tol = 0.01)
}
\arguments{
\item{x}{exponentially tilted positive stable observation vector}

\item{alpha}{positive stable index (0 < alpha < 1)}

\item{delta}{Hougaard scale parameter}

\item{theta}{Hougaard exponentially tilting parameter}

\item{return_log}{(logical) should log-likelihood be returned}

\item{tol}{tolerance for alpha parameter (bounding away from 0 and 1)}
}
\value{
exponentially tilted positive stable density
}
\description{
R implementation of density function for exponentially tilted PS(alpha)
}
