% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{dps_oneR}
\alias{dps_oneR}
\title{See Stephenson (2009) equation (2) integral for single observation}
\usage{
dps_oneR(x, alpha, return_log = TRUE)
}
\description{
See Stephenson (2009) equation (2) integral for single observation
}
