% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{exact_lap}
\alias{exact_lap}
\title{Exact Laplace transform of Hougaard distribution}
\usage{
exact_lap(t, alpha, delta, theta)
}
\description{
Exact Laplace transform of Hougaard distribution
}
