% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{exact_mean}
\alias{exact_mean}
\title{Exact mean of Hougaard distribution (for theta > 0)}
\usage{
exact_mean(alpha, delta, theta)
}
\description{
Exact mean of Hougaard distribution (for theta > 0)
}
