% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_measures.R
\name{fixed_chi_exact}
\alias{fixed_chi_exact}
\title{Calculate theoretical chi_u in Gaussian density basis case}
\usage{
fixed_chi_exact(nq, qlim, alpha, theta, obs_coord, knot_coord, kern_bw,
  lK = NULL, us = NULL)
}
\arguments{
\item{nq}{number of quantiles at which to calculate chi_u. This is only used
if \code{us} is not passed.}

\item{qlim}{Range of u over which chi_u should be calculated.}

\item{alpha}{Positive stable index}

\item{theta}{Hougaard exponential tilting parameter}

\item{obs_coord}{Observation coordinate matrix/vector defining two locations
for which chi_u should be calculated.}

\item{knot_coord}{Gaussian density basis function knot coordinates}

\item{kern_bw}{Gaussian density basis kernel bandwidth (sd)}

\item{lK}{2 x nbasis matrix of log-transformed, normalized, log-Gaussian
density basis functions}

\item{us}{vector of u values to calculate chi_u for}
}
\value{
chi_u, chibar_u and u
}
\description{
Calculate theoretical chi_u in Gaussian density basis case
}
