% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{gamf}
\alias{gamf}
\title{Gamma function, which is used in parameterization of positive stable
distribution in the \code{stabledist} package}
\usage{
gamf(alpha)
}
\description{
Gamma function, which is used in parameterization of positive stable
distribution in the \code{stabledist} package
}
