% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{get_gp}
\alias{get_gp}
\title{Subtract known mean from Gaussian process to get residual (mean zero)
process.}
\usage{
get_gp(mpar, X, betas, nrep, log.link = FALSE)
}
\arguments{
\item{mpar}{Matrix containing gaussian process copies in all rows}

\item{X}{design matrix of known covariates for GP mean}

\item{betas}{linear coefficients for X matrix}

\item{nrep}{number of rows in mpar}

\item{log.link}{(logical) was a log-link used to construct the mpar GP}
}
\description{
Subtract known mean from Gaussian process to get residual (mean zero)
process.
}
