% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{gev2lsm}
\alias{gev2lsm}
\title{Transform marginal distribution from log-stablemix to gev}
\usage{
gev2lsm(y, mu, sigma, xi, qfuns)
}
\arguments{
\item{y}{(matrix) gev scale data to transform. Columns: locations,
rows: replicates.}

\item{mu}{GEV location parameter}

\item{sigma}{GEV scale parameter}

\item{xi}{GEV shape parameter}

\item{qfuns}{(list of log-stablemix quantile functions) for each column of y.
See \code{create_qlstabmix}}
}
\value{
Matrix of same dimension as \code{y} of transformed GEV data to
log-stable mixture scale
}
\description{
Transform from gev to log-stablemix.
Need to run create_qlstabmix first.
}
