% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{gev2lsm_all}
\alias{gev2lsm_all}
\title{Wrapper for create_qlstabmix and gev2lsm}
\usage{
gev2lsm_all(y, alpha, theta, lK, mu, sigma, xi)
}
\arguments{
\item{y}{(matrix) gev scale data to transform. Columns: locations,
rows: replicates.}

\item{alpha}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{theta}{stable theta parameter (exponential tilting parameter). This
parameter controls the tail dependence class. \code{theta} = 0 gives}

\item{mu}{GEV location parameter}

\item{sigma}{GEV scale parameter}

\item{xi}{GEV shape parameter}
}
\value{
Matrix of same dimension as \code{y} of transformed GEV data to
log-stable mixture scale
}
\description{
All-in-one transformation from GEV to log-stabmix
}
