% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{gev2lsm_w_na}
\alias{gev2lsm_w_na}
\title{Convert GEV observations to log-stable mixture scale, but leave NA alone}
\usage{
gev2lsm_w_na(y, mu, sigma, xi, qfuns)
}
\arguments{
\item{y}{GEV scale observation matrix (each column corresponds to a different
location)}

\item{mu}{GEV location parameter matrix}

\item{sigma}{GEV scale parameter matrix}

\item{xi}{GEV shape parameter matrix}

\item{qfuns}{Quantile functions for log-stable mixture distribution. Should
be equal in length to the number of columns of \code{y} matrix.}
}
\value{
Marginally transformed matrix of y observations (to log-stable
mixture scale).
}
\description{
Convert GEV observations to log-stable mixture scale, but leave NA alone
}
