% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{laplace_mc}
\alias{laplace_mc}
\title{Monte Carlo estimate of Laplace transform of Hougaard distribution}
\usage{
laplace_mc(t, alpha, delta, theta, n = 10000)
}
\description{
used for a sanity check to compare against known Laplace transform
}
