% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{llgp}
\alias{llgp}
\title{GP Log-likelihood wrapper (RandomFields)}
\usage{
llgp(gp, s, gvar, gscl)
}
\arguments{
\item{gp}{Matrix of Gaussian process realizations (see RandomFields for
format)}

\item{s}{nloc x 2 matrix of observation location coordinates}

\item{gvar}{GP variance}

\item{gscl}{GP scale/range}
}
\value{
log-likelihood of GP observations
}
\description{
RandomFields wrapper to calculate log-likelihood of GP with
exponential covariance function
}
