% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{llikM_w_na}
\alias{llikM_w_na}
\title{Calculate observation model log-likelihood excluding NAs}
\usage{
llikM_w_na(lz, y, lA, lK, alpha, theta, mu, sigma, xi, tol = 0.02)
}
\arguments{
\item{lz}{matrix nrep x nloc. log-stable mixture observations}

\item{y}{matrix nrep x nloc. log-stable mixture observations transformed to
GEV scale with location, scale, and shape given by mu, sigma, and xi}

\item{lA}{(n by number of obs_coord) matrix of 1/alpha-norm scaled
kernel functions at each observation location. Transformed to log
scale.}

\item{lK}{(number of obs_coord by number of knot_coord) matrix of
log-kernel functions at each observation location.}

\item{alpha}{Positive stable index}

\item{theta}{Hougaard exponential tilting parameter}

\item{mu}{GEV location parameter matrix}

\item{sigma}{GEV scale parameter matrix}

\item{xi}{GEV shape parameter matrix}

\item{tol}{tolerance bounding alpha away from zero and one}
}
\value{
observation model log-likelihood
}
\description{
Calculate observation model log-likelihood excluding NAs
}
