% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{llKbasis}
\alias{llKbasis}
\title{Log-prior function for log-Gaussian process kernels}
\usage{
llKbasis(lK, s, gvar, gscl)
}
\arguments{
\item{lK}{log-Gaussian process kernels which sum to 1 at each location after
exponentiation. Matrix dimensions: nloc x nbasis}

\item{s}{observation location matrix (nloc x 2)}

\item{gvar}{GP variance}

\item{gscl}{GP scale/range}
}
\value{
log-prior for kernel functions.
}
\description{
Log-prior function for log-Gaussian process kernels
}
