% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_measures.R
\name{lnorm_cond_chi_exact}
\alias{lnorm_cond_chi_exact}
\title{Calculate conditional chi_u conditionally on basis functions}
\usage{
lnorm_cond_chi_exact(nq, qlim, alpha, theta, obs_coord, L, gvar, gscl,
  lK = NULL, us = NULL)
}
\arguments{
\item{nq}{number of quantiles at which to calculate chi_u. This is only used
if \code{us} is not passed.}

\item{qlim}{Range of u over which chi_u should be calculated.}

\item{alpha}{Positive stable index}

\item{theta}{Hougaard exponential tilting parameter}

\item{obs_coord}{Observation coordinate matrix/vector defining two locations
for which chi_u should be calculated.}

\item{gvar}{GP variance}

\item{lK}{2 x nbasis matrix of log-transformed, normalized, basis functions}

\item{us}{vector of u values to calculate chi_u for}

\item{gvar}{GP exponential cov. range/scale}
}
\value{
chi_u, chibar_u and u
}
\description{
Calculate conditional chi_u conditionally on basis functions
}
