% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_simulation_low_level.R
\name{lnorm_cond_sim_low_all}
\alias{lnorm_cond_sim_low_all}
\title{Make conditional simulations of response conditional on low-level processes (log-normal basis)}
\usage{
lnorm_cond_sim_low_all(out, pred_coords, thin_int = 1, sim_y = TRUE)
}
\arguments{
\item{out}{MCMC output object}

\item{pred_coords}{matrix of prediction coordinates}

\item{thin_int}{Thinning interval}

\item{sim_y}{(logical) Should the response be simulated at prediction locations}
}
\description{
Make conditional simulations of response conditional on low-level processes (log-normal basis)
}
