% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{lps}
\alias{lps}
\title{Log-likelihood of positive stable distribution}
\usage{
lps(x, alpha, return_log = T, tol = 0.01)
}
\arguments{
\item{x}{positive stable observation vector}

\item{alpha}{positive stable index (0 < alpha < 1)}

\item{return_log}{(logical) should log-likelihood be returned?}

\item{tol}{numeric tolerance for alpha parameter (away from zero and 1)}
}
\value{
log-likelihood of positive stable distribution
}
\description{
Log-likelihood of positive stable distribution
}
\examples{
a <- rps(100, 0.5)
lps(a, 0.5)
}
