% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{lpst_gp_cov_par}
\alias{lpst_gp_cov_par}
\title{Log-full conditional for GP exp. covariance parameters}
\usage{
lpst_gp_cov_par(gp, s, gvar, gscl, rng_max = NULL)
}
\arguments{
\item{gp}{matrix of Gaussian process realizations (See RandomFields for form)}

\item{s}{nloc x 2 matrix of observation location coordinates}

\item{gvar}{GP variance}

\item{gscl}{GP scale/range}

\item{rng_max}{optionally define prior for range parameter in terms of the
range of the observation coordinates.}
}
\description{
Log-full conditional for GP exp. covariance parameters
}
