% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{lpst_gvar}
\alias{lpst_gvar}
\title{Log-full conditional for log-GP basis variance parameter}
\usage{
lpst_gvar(gvar, s, lK, gscl)
}
\arguments{
\item{gvar}{GP variance}

\item{s}{observation location matrix (nloc x 2)}

\item{lK}{log-Gaussian process kernels which sum to 1 at each location after
exponentiation. Matrix dimensions: nloc x nbasis}

\item{gscl}{GP scale/range}
}
\description{
Log-full conditional for log-GP basis variance parameter
}
