% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{lpst_mpar_mean}
\alias{lpst_mpar_mean}
\title{Log-full conditional for marginal GEV parameter GP mean function parameters}
\usage{
lpst_mpar_mean(beta, lz, y, lA, lK, alpha, theta, mu, sigma, xi,
  is.xi = FALSE)
}
\arguments{
\item{beta}{GP mean function coefficients in X beta}

\item{lz}{(n-rep by nloc) matrix of log-transformed stable-mixture
observations}

\item{y}{(n-rep by nloc) matrix of stable-mixture observations on GEV scale}

\item{lA}{(n by number of obs_coord) matrix of 1/alpha-norm scaled
kernel functions at each observation location.}

\item{lK}{(nloc by nbasis) matrix of log-transformed sum constrained kernel
functions}

\item{alpha}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{theta}{stable theta parameter (exponential tilting parameter). This
parameter controls the tail dependence class. \code{theta} = 0 gives
asymptotic dependence, and \code{theta} > 0 gives asymptotic independence.}

\item{mu}{GEV location matrix}

\item{sigma}{GEV scale matrix}

\item{xi}{GEV shape matrix}

\item{is.xi}{(logical) is modeled GP for the shape parameter?}
}
\description{
Log-full conditional for marginal GEV parameter GP mean function parameters
}
