% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{lpst_theta}
\alias{lpst_theta}
\title{Full-conditional for theta parameter}
\usage{
lpst_theta(theta, lz, y, lB, lA, lK, alpha, mu, sigma, xi)
}
\arguments{
\item{theta}{stable theta parameter (exponential tilting parameter). This
parameter controls the tail dependence class. \code{theta} = 0 gives
asymptotic dependence, and \code{theta} > 0 gives asymptotic independence.}

\item{lz}{(n-rep by nloc) matrix of log-transformed stable-mixture
observations}

\item{y}{(n-rep by nloc) matrix of stable-mixture observations on GEV scale}

\item{lB}{matrix of log-basis scaling factor matrix (e.g. from Hougaard
distribution) dimension: n basis x n independent replicates}

\item{lA}{(n by number of obs_coord) matrix of 1/alpha-norm scaled
kernel functions at each observation location.}

\item{lK}{(nloc by nbasis) matrix of log-transformed sum constrained kernel
functions}

\item{alpha}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{mu}{GEV location matrix}

\item{sigma}{GEV scale matrix}

\item{xi}{GEV shape matrix}
}
\value{
log-full conditional for theta
}
\description{
Full-conditional for theta parameter
}
