% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{lsm2gev}
\alias{lsm2gev}
\title{Transform marginal distribution from log-stablemix to gev}
\usage{
lsm2gev(lz, alpha, theta, lK, mu, sigma, xi)
}
\arguments{
\item{lz}{log-stable mixture observation matrix (rows: independent replicates
columns: observation locations)}

\item{alpha}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{theta}{stable theta parameter (exponential tilting parameter). This
parameter controls the tail dependence class. \code{theta} = 0 gives
asymptotic dependence, and \code{theta} > 0 gives asymptotic independence.}

\item{lK}{log-kernel basis matrix (rows: locations, columns: different
kernel functions)}

\item{mu}{GEV location vector (one value per observation location)}

\item{sigma}{GEV scale vector (one value per observation location)}

\item{xi}{GEV shape vector (one value per observation location)}
}
\value{
Transformed matrix of observations on GEV scale
}
\description{
Transform marginal distribution from log-stablemix to gev
}
\examples{
alpha <- 0.9                # positive stable index
theta <- 0.01               # tilting parameter
tau <- 0.1                  # kernel bandwidth
n <- 30                     # number of replicates (years)
nloc <- 10^2                # number of spatial locations
L <- 25                     # number of random effects per year
s <- as.matrix(expand.grid(seq(0.1,0.9, l = sqrt(nloc)),
                           seq(0.1,0.9, l = sqrt(nloc))))
v <- as.matrix(expand.grid(seq(0, 1, l = sqrt(L)),
                           seq(0, 1, l = sqrt(L))))
lZ <- rstabmix(n, s, v, nbasis = L, alpha = alpha,
               delta = alpha, theta = theta,
               kern_bw = tau, type = "smith",
               gvar = gvar, gscl =gscl)
mu <- sigma <- matrix(1, nrow = n, ncol = nloc)
xi <- matrix(0, nrow = n, ncol = nloc)
y <- lsm2gev(lZ, alpha, theta, lK, mu, sigma, xi)
}
