% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{make_vnames}
\alias{make_vnames}
\title{Add prefix to a variable name (helper)}
\usage{
make_vnames(x, prefix)
}
\arguments{
\item{x}{variable name}

\item{prefix}{string to add before variable name with underscore}
}
\value{
prefixed string
}
\description{
If variable name is missing, titled "(Intercept)"
}
