% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{mc_mean}
\alias{mc_mean}
\title{Monte Carlo estimate of Hougaard mean (for theta > 0)}
\usage{
mc_mean(alpha, delta, theta, n = 10000)
}
\description{
Monte Carlo estimate of Hougaard mean (for theta > 0)
}
