% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_lognormal_basis.R
\name{mcmc_lnorm_basis}
\alias{mcmc_lnorm_basis}
\title{Metropolis-Hastings algorithm for log-Gaussian process basis stable mixture model}
\usage{
mcmc_lnorm_basis(n_iter, obs_coord, y, lB, lK, init, tune_var, data,
  mar_gp_par, mar_gp_init = NULL, mar_gp_which = NULL, loc = ~1,
  scale = ~1, shape = ~1, opt_rt = 0.4, select_lat = NULL,
  thin_int = 1, alpha_tol = 0.07, clusters = NULL,
  infer_general_theta = FALSE, parallelize = FALSE)
}
\arguments{
\item{n_iter}{number of MCMC iterations/posterior samples to generate}

\item{obs_coord}{n-location x 2 matrix of observation location coordinates}

\item{y}{n-replicate x n-location matrix of observations with GEV margins
to which a stable-mixture model will be fit.}

\item{lB}{n-basis x n-replicate matrix of initial values for log-transformed
random kernel scaling factors (assumed to follow a Hougaard distribution).}

\item{lK}{n-location x n-basis matrix of initial values for normalized
(sum to 1) log-Gaussian process prior kernel functions, which are transformed
to the log-scale after normalization.}

\item{init}{named list of initial values for model parameters. Should contain
alpha, theta, tau (kernel bw), loc (mean function betas for GEV location),
scale (mean function betas for log(GEV scale)), shape.}

\item{tune_var}{named list of tuning proposal variance parameters containing
alpha, theta, tau (kernel bandwidth) loc, scale, shape (each of which should
match the beta coefficient vector dimension in e.g. loc = X %*% beta),
lB, and gps (which is itslef a list with names gvar, gscl, and gp) for GEV
marginal parameter Gaussian processes.}

\item{data}{data.frame containing covariates with which to model GEV
marginal parameter means.}

\item{mar_gp_par}{Named list containing initial values for GEV marginal
parameter Gaussian process hyper parameter initial values. E.g.,
list(mu = list(gvar = 0.01, gscl = 1/3),
sigma = list(gvar = 0.01, gscl = 1/3),
xi =list(gvar = 0.01, gscl = 1/3))}

\item{mar_gp_init}{(Optional) list of initial values for GEV location,
scale, and shape GP residual processes (i.e. with mean zero).}

\item{mar_gp_which}{vector containing a subset of c("mu","sigma","xi")
indicating which GEV marginal parameters should be modeled as spatial
Gaussian processes, and which should be held constant in space. Note: if
"sigma" is specified, it will be modeled as a log-Gaussian process to
satisfy the positivity constraint.}

\item{loc}{R formula for (GP mean function of, depending) GEV location
parameter. Any specified covariates should be present in 'data'}

\item{scale}{R formula for (GP mean function of, depending) GEV scale
parameter. Any specified covariates should be present in 'data'}

\item{shape}{R formula for (GP mean function of, depending) GEV shape
parameter. Any specified covariates should be present in the 'data'}

\item{opt_rt}{Target acceptance rate for sampler (applies to all parameters)
Should be a number between 0 and 1.}

\item{select_lat}{If only a subset of the latent parameters should be
returned (e.g. for the first 5 locations), pass a vector of indices (e.g. 1:5)}

\item{thin_int}{Thinning interval if there are memory constraints for
mcmc sample storage.}

\item{alpha_tol}{(optional) tolerance to bound alpha parameter away from 0
and 1.}

\item{clusters}{(optional) list of indices for homogeneous subregions of
observation locations (should match row numbers of observation location
matrix). This blocking scheme is used to improve the efficiency of the
Gaussian process updates.}

\item{infer_general_theta}{(not currently implemented)}

\item{parallelize}{(logical) If multiple cores are available, can parallelize
over the independent replicates of the process when updating the random
basis scaling factors belonging to different replicates (e.g., years)}

\item{tau_tol}{(optional) kernel bandwidth tolerance to bound bandwidth
away from zero, which might cause kernel degeneracy and numerical instability.}
}
\value{
list containing
  smp_par: posterior samples of model parameters
  smp_lB: posterior samples of basis scaling factors (on log scale)
  smp_lK: posterior samples of normalized log-Gaussian process basis
          functions (on log scale)
  smp_mu_gp: location parameter (mean-zero) residual gaussian process
  smp_sigma_gp: log(scale parameter) (mean-zero) residual gaussian process
  smp_xi_gp: shape parameter (mean-zero) residual gaussian process
  smp_mu_gvar: GEV location parameter GP variance parameter posterior samples
  smp_sigma_gvar: GEV log(scale parameter) GP variance parameter posterior samples
  smp_xi_gvar: GEV shape parameter GP variance parameter posterior samples
  smp_mu_gscl: GEV location parameter GP location parameter posterior samples
  smp_sigma_gvar: GEV log(scale parameter) GP location parameter posterior samples
  acpt_rts: matrix of acceptance rates for sampled parameters
  acpt_rts_loc: matrix of acceptance rates for GEV location parameter GPs
  acpt_rts_scale: matrix of acceptance rates for GEV log(scale) parameter GPs
  acpt_rts_shape: matrix of acceptance rates for GEV shape parameter GPs
  acpt_rts_lB: matrix of acceptance rates for random basis scaling factors
  acpt_rts_lK: matrix of acceptance rates for random basis function terms
  prop_var: proposal variances for model parameter updates
  loc_prop_var: proposal variances for GEV location parameter
  scale_prop_var: proposal variances for GEV log(scale) parameter
  shape_prop_var: proposal variances for GEV shape parameter
  clusters: observation location cluster block definitions for efficient
  mcmc sampling
}
\description{
Main Metropolis-Hastings loop to generate posterior samples for
stable-mixture model when using log-Gaussian process kernel function basis.
}
