% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_measures.R
\name{prepare_fixed_chis}
\alias{prepare_fixed_chis}
\title{Wrapper for fixed_chi_exact to calculate chi_u by iterating over several parameters}
\usage{
prepare_fixed_chis(nq, qlim, obs_coord, knot_coord, kern_bw, alphas,
  thetas)
}
\arguments{
\item{nq}{number of quantiles at which to calculate chi_u. This is only used
if \code{us} is not passed.}

\item{qlim}{Range of u over which chi_u should be calculated.}

\item{obs_coord}{Observation coordinate matrix/vector defining two locations
for which chi_u should be calculated.}

\item{knot_coord}{Gaussian density basis function knot coordinates}

\item{kern_bw}{Gaussian density basis kernel bandwidth (sd)}

\item{alphas}{vector of positive stable indices for which to calculate chi_u}

\item{thetas}{Hougaard exponential tilting parameters for which to calculate
chi_u}
}
\value{
data.frames of chi_u and chibar_u for different parameter settings
}
\description{
Wrapper for fixed_chi_exact to calculate chi_u by iterating over several parameters
}
