% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_measures.R
\name{prepare_lnorm_chis}
\alias{prepare_lnorm_chis}
\title{Monte Carlo estimate of chi_u as a function of distance for log-GP basis}
\usage{
prepare_lnorm_chis(nq, qlim, obs_coord, L, gvar, gscl, alphas, thetas,
  nmc = 100)
}
\arguments{
\item{nq}{number of quantiles at which to calculate chi_u. This is only used
if \code{us} is not passed.}

\item{qlim}{Range of u over which chi_u should be calculated.}

\item{obs_coord}{Observation coordinate matrix/vector defining two locations
for which chi_u should be calculated.}

\item{L}{number of basis functions}

\item{gvar}{GP variance}

\item{alphas}{vector of positive stable indices for which to calculate chi_u}

\item{thetas}{Hougaard exponential tilting parameters for which to calculate
chi_u}

\item{nmc}{Number of Monte Carlo replicates to use.}

\item{lK}{2 x nbasis matrix of log-transformed, normalized, basis functions}

\item{gvar}{GP exponential cov. range/scale}
}
\value{
data.frames of chi_u and chibar_u for different parameter settings

data.frames of chi_u and chibar_u for different parameter settings
}
\description{
Wrapper for lnorm_cond_chi_exact to calculate chi_u by iterating
over several parameters. Monte Carlo is used to marginalize over basis
functions.
}
