% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{pstabmix}
\alias{pstabmix}
\title{Stable mixture CDF}
\usage{
pstabmix(z, alpha, theta, basis = NULL, kern_bw = NULL,
  obs_coord = NULL, knot_coord = NULL, logp = FALSE)
}
\arguments{
\item{z}{vector of quantiles}

\item{alpha}{stable alpha parameter}

\item{theta}{stable theta parameter (exponential decay)}

\item{basis}{Matrix of basis functions at observation locations}

\item{kern_bw}{Kernel bandwidth parameter (if using Gaussian density basis)}

\item{obs_coord}{Matrix of observation coordinates}

\item{knot_coord}{Matrix of knot coordinates (if using Gaussian density basis)}

\item{logp}{(logical) Should the log-cdf be returned?}
}
\description{
Stable mixture CDF
}
