% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evd_functions.R
\name{qevdM}
\alias{qevdM}
\title{R wrapper for C function to calculate GEV quantile function}
\usage{
qevdM(p, loc, scale, shape)
}
\arguments{
\item{p}{matrix of probabilities for which to calculate quantiles}

\item{loc}{matrix of GEV location parameters}

\item{scale}{matrix of GEV scale parameters}

\item{shape}{matrix of GEV shape parameters}

\item{lower.tail}{(logical) should lower tail be returned?}
}
\value{
GEV quantiles
}
\description{
R wrapper for C function to calculate GEV quantile function
}
