% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{qlstabmix_rf}
\alias{qlstabmix_rf}
\title{log-stable mixture observation model quantile estimation}
\usage{
qlstabmix_rf(p, alpha, theta, basis)
}
\arguments{
\item{p}{desired quantile (0 < p < 1)}

\item{alpha}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{theta}{stable theta parameter (exponential tilting parameter). This
parameter controls the tail dependence class. \code{theta} = 0 gives
asymptotic dependence, and \code{theta} > 0 gives asymptotic independence.}

\item{basis}{Vector of sum=1 constrained basis functions at single
observation location.}
}
\description{
Numerical inversion of log-stablemix observation model CDF using
 root finding for a single location
}
