% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{qstabmix}
\alias{qstabmix}
\title{Stable mixture quantile function}
\usage{
qstabmix(p, alpha, theta, basis)
}
\arguments{
\item{p}{vector of probabilities}

\item{alpha}{stable alpha parameter}

\item{theta}{stable theta parameter (exponential decay)}

\item{basis}{Matrix of basis functions at observation locations}
}
\description{
Stable mixture quantile function
}
\examples{
# Not run
n <- 100
L <- 10
tau <- 1
v <- seq(0,1, l = L)
s <- seq(0.1, 0.9, l = n)
K <- make_kernel(s, L, "smith", knot_coord = v, kern_bw = tau) # Random basis functions
matplot(s, K, type = "l")
basis <- K[1,]
qout <- qstabmix(0.5, 0.5, 0.1, 0, basis)
pstabmix(qout$root, 0.5, 0.1, 0, basis)
}
