% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{qstabmix_spline}
\alias{qstabmix_spline}
\title{Calculate quantile function for stable-mixture observation model}
\usage{
qstabmix_spline(alpha, theta, lK, grd_len = 100)
}
\arguments{
\item{alpha}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{theta}{stable theta parameter (exponential tilting parameter). This
parameter controls the tail dependence class. \code{theta} = 0 gives
asymptotic dependence, and \code{theta} > 0 gives asymptotic independence.}

\item{lK}{matrix of log-basis functions evaluated at observation locations.
nrow: number of observation locations. ncol: number of basis functions.}

\item{grd_len}{number of design points when defining grid for numerical
inversion of CDF. Grids of 100 or more have been consistent with root-finding
inversion for stable-mixture cdfs.}
}
\value{
R function that takes a single argument (probability 0 < p < 1) and returns
corresponding stable-mixture observation model quantile function
}
\description{
Calculate quantile function by numerically inverting the CDF.
Since the marginal distributions are assumed to be the same across years but
vary by spatial location, storing the quantile functions can be dramatically
faster than performing root-finding for each observation when the number of
years is large.
}
