% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{rexpstab_one}
\alias{rexpstab_one}
\title{Simulate single Hougaard RV using Devroye (2009) in reparameterized form}
\usage{
rexpstab_one(alpha, lambda)
}
\arguments{
\item{alpha}{positive stable index (0 < alpha < 1)}

\item{lambda}{= theta * (alpha/delta) ^ (1 / alpha)}
}
\value{
single realization from Hougaard dist'n
}
\description{
Apply the Method of Luc Devroye (2009) for exponentially
tilted stable distributions from the paper
`Random Variate Generation for Exponentially and
Polynomially Tilted Stable Distributions'. This method applies only to
the theta > 0 case. Here parameterized using the Devroye (2009) lambda
parameter. See section 9 of Devroye (2009). This algorithm is useful for
large theta.
}
