% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{rgp}
\alias{rgp}
\title{Simulate Gaussian process}
\usage{
rgp(s, gvar, gscl)
}
\arguments{
\item{s}{(nloc x 2 dimensional matrix) of locations at which to simulate GP}

\item{gvar}{GP variance parameter}

\item{gscl}{GP scale/range parameter}
}
\description{
Simulate a Gaussian process (GP)with exponential covariance
function using RandomFields pacakge.
}
\examples{
nloc = 100
s <- as.matrix(expand.grid(seq(0.1,0.9, l = sqrt(nloc)),
                           seq(0.1,0.9, l = sqrt(nloc))))
Z <- rgp(s, 1, 1)
}
