% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{rhexpstab}
\alias{rhexpstab}
\title{Simulate from Hougaard distribution}
\usage{
rhexpstab(n, alpha, delta, theta)
}
\arguments{
\item{n}{number of realizations to simulate}

\item{alpha}{positive stable index (0 < alpha < 1)}

\item{delta}{Hougaard scale parameter}

\item{theta}{Hougaard exponential tilting parameter}
}
\value{
n realizations from Hougaard(alpha, delta, theta)
}
\description{
Simulate from Hougaard distribution
}
