% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{rnorm_prop}
\alias{rnorm_prop}
\title{Random walk Normal proposal}
\usage{
rnorm_prop(cur, tune_var)
}
\arguments{
\item{cur}{current mcmc sample to center proposal at}

\item{tune_var}{tuning variance for mcmc proposal}
}
\value{
random walk Normal proposal
}
\description{
Random walk Normal proposal
}
