% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{rpostpred}
\alias{rpostpred}
\title{Make conditional draws from stable-mixture dist'n}
\usage{
rpostpred(n, obs_coord, alpha, lK, lB, sub_index = NULL,
  return_logZ = TRUE)
}
\arguments{
\item{n}{number of replicates to simulate}

\item{obs_coord}{ncoord x 2 dimensional matrix of observation coordinates}

\item{alpha}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{lK}{ncoord x nbasis dimensional matrix of log-transformed normalized
kernels at each observation location. This should come from mcmc posterior
draws.}

\item{lB}{nreplicate x nyear matrix of log-transformed basis scaling factors.
This should come from mcmc posterior draws.}

\item{sub_index}{Indices of observation locations if only a subset of
locations are desired}

\item{return_logZ}{(logical) should draws of stable-mixture variates be
returned on the log-scale?}
}
\description{
Make a draw from stable-mixture conditional on the basis
functions, scaling factors, etc.
}
