% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{rps}
\alias{rps}
\title{Random positive stable distribution simulation wrapper of stabledist}
\usage{
rps(n, alpha)
}
\arguments{
\item{n}{number of realizations to simulate}

\item{alpha}{positive stable index (0 < alpha < 1)}
}
\value{
realizations from positive stable distribution
}
\description{
Random positive stable distribution simulation wrapper of stabledist
}
\examples{
a <- rps(100, 0.5)
}
