% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{rstabmix}
\alias{rstabmix}
\title{Simulate from stable-mixture model}
\usage{
rstabmix(n, obs_coord, knot_coord, nbasis, alpha, delta, theta,
  kern_bw = NULL, gvar = NULL, gscl = NULL, type = c("smith",
  "br"), return_all = FALSE, return_logZ = TRUE)
}
\arguments{
\item{n}{number of independent replicates to generate (e.g., years)}

\item{obs_coord}{matrix of observation coordinates. The rows correspond
to different locations, and the columns correspond to the dimensions of
those coordinates.}

\item{knot_coord}{matrix of knot locations. The rows correspond to
different locations. Needed only if \code{type} = 'smith'.}

\item{nbasis}{number of kernel (basis) functions to generate. Kernels will
sum to one across all nbasis at each location.}

\item{alpha}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{delta}{stable delta parameter (scale): \code{delta} > 0.
Elsewhere, this is fixed equal to \code{alpha}.}

\item{theta}{stable theta parameter (exponential tilting parameter). This
parameter controls the tail dependence class. \code{theta} = 0 gives
asymptotic dependence, and \code{theta} > 0 gives asymptotic independence.}

\item{kern_bw}{Kernel bandwidth (sd) for 'smith' kernel bases}

\item{gvar}{Log-Gaussian process variance term ('brown-resnick'). Needed
only if \code{type} = 'br'}

\item{gscl}{Log-Gaussian process scale/range term ('brown-resnick'). Needed
only if \code{type} = 'br'}

\item{type}{(string) type of kernel function to create. Should be one of
either c("smith", "br"). For fixed, Gaussian density
kernels, use "smith". For a log-Gaussian process kernels, use "br",
which stands for Brown-Resnick.}

\item{return_all}{(logical) TRUE: just return stable mixture process,
FALSE: return components}

\item{return_logZ}{(logical) should return stable mixture realizations be
returned on the log-scale?}
}
\value{
list containing:
        lZ: (n by number of obs_coord) matrix of (log) stable mixture
        process observations.
        lA: (n by number of obs_coord) matrix of 1/alpha-norm scaled
        kernel functions at each observation location.
        lK: (number of obs_coord by number of knot_coord) matrix of
        log-kernel functions at each observation location.
        lB: (number of bases by n) stable random effects (i.e. kernel scaling
        terms).
}
\description{
Simulate n realizations from stable-mixture model
}
\examples{
n <- 3                    # Number of independent replicates (years)
L <-30                    # Number of random effects per year
alpha <- 0.1             # Stable alpha
theta <- 0.001            # Stable exp tilt
tau <- 0.5                  # Gauss kernel bw (sd)
v <- seq(0, 10, l = L)    # Kernel knots
s <-seq(3,8, l = 1000)     # Spatial locations
Z <- rstabmix(n, obs_coord = s, knot_coord = v,
              nbasis = L, alpha = alpha, delta = alpha,
              theta = theta, kern_bw = tau)
matplot(s, exp(t(Z)), type = "l")
}
