% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_mixture.R
\name{sample_marpar_gp}
\alias{sample_marpar_gp}
\title{Simulate marginal GEV parameter GP processes}
\usage{
sample_marpar_gp(formula, data, betas, nrep, s, cov_par = NULL,
  log.link = FALSE, return_all = FALSE)
}
\arguments{
\item{formula}{R formula for GP mean function using covariates in \code{data}}

\item{data}{data.frame containing covariates specified in formula}

\item{betas}{vector of coefficients with which to linearly combine covariates}

\item{nrep}{number of data replicates of stable mixture process with common
margins across time.}

\item{s}{nloc x 2 dimensional matrix of observation locations}

\item{cov_par}{list of GP covariance function parameters. Must contain
named gvar and gscl elements. See \code{rgp}.}

\item{log.link}{(logical) should a log link function be used. Useful for GEV
scale parameter, which must be positive.}

\item{return_all}{(logical) should the mean-zero residual process (to the
known mean) be returned in addition to the GP with non-zero mean?}
}
\description{
Simulate a Gaussian process with exponential covariance
and known mean that may depend on covariates. Used for generating
marginal GEV parameters.
}
