% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tilted_alpha_stable.R
\name{sinc}
\alias{sinc}
\title{sinc function}
\usage{
sinc(x, tol = 1e-10)
}
\arguments{
\item{x}{sinc argument}

\item{tol}{tolerance to bound from zero}
}
\description{
Apply the Method of Luc Devroye (2009) for exponentially
             tilted stable distributions
}
\examples{
x <- seq(1e-20, 0.01, l= 1000)
s <- sapply(x, sinc)
plot(x, s, type = "l")
}
