% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{update_lB}
\alias{update_lB}
\title{Metropolis update for random scaling factors (supplied on log-scale)}
\usage{
update_lB(lB_cur, lzsub, ysub, lK_cur, alpha_cur, theta_cur, mu_cur,
  sigma_cur, xi_cur, pos, next_pos, acpt_lB, lB_prop_var)
}
\arguments{
\item{lB_cur}{Current mcmc samples of log-scaling factor for a single year}

\item{lzsub}{log-stable mixture observations for the corresponding year to
the supplied lB_cur vector}

\item{ysub}{GEV scale observations after marginal transformations of lzsub}

\item{lK_cur}{(nloc by nbasis) matrix of log-kernel functions for current
mcmc iteration}

\item{alpha_cur}{stable alpha parameter: 0 < \code{alpha} < 1. Controls the
tradeoff between all signal (alpha = 0) and all nugget (alpha = 1).}

\item{theta_cur}{stable theta parameter (exponential tilting parameter). This
parameter controls the tail dependence class. \code{theta} = 0 gives
asymptotic dependence, and \code{theta} > 0 gives asymptotic independence.}

\item{mu_cur}{GEV location matrix for current mcmc iteration}

\item{sigma_cur}{GEV scale matrix for current mcmc iteration}

\item{xi_cur}{GEV shape matrix for current mcmc iteration}

\item{pos}{index in adaptive variance updates matrix}

\item{next_pos}{next index in adaptive variance updates matrix}

\item{acpt_lB}{acceptance matrix of (1: accept) and (0: reject) for lB
scaling factors. Used for determining adaptive proposal updates.}

\item{lB_prop_var}{proposal variance vector for the lB scaling factors}
}
\description{
Metropolis update for random scaling factors (supplied on log-scale)
}
