% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{update_var}
\alias{update_var}
\title{Spring function for automatic proposal variance tuning update}
\usage{
update_var(cur_var, acpt_rt, opt_rt, gamma1)
}
\arguments{
\item{cur_var}{current proposal variance}

\item{acpt_rt}{current acceptance rate (between 0 and 1)}

\item{opt_rt}{target acceptance rate (between 0 and 1)}

\item{gamma1}{spring tuning parameter}
}
\value{
new proposal variance
}
\description{
Spring function for automatic proposal variance tuning update
}
